feat: 전시회 등록 페이지와 에디터 연결
- CSS 일부 수정
- Next/Back 버튼을 이용한 내부페이지 이동
