+++ client/pages/auction/[id].tsx
@@ -8,13 +8,7 @@ import Layout from '@components/common/Layout';
 import ItemDetail from '@components/Auction/ItemDetail';
 import { GlobalStore } from '../../store/GlobalStore';
 
-const DUMMY_DATA: Artwork = {
-    id: 1,
-    imagePath:
-        'https://d7hftxdivxxvm.cloudfront.net/?resize_to=fit&width=210&height=276&quality=80&src=https%3A%2F%2Fd32dm0rphc51dk.cloudfront.net%2FVju3jVJD5yaSEb1vTQbA1w%2Flarge.jpg',
-};
-
-const AuctionDetailPage = ({ artwork = DUMMY_DATA }: { artwork: Artwork }) => {
+const AuctionDetailPage = ({ artwork }: { artwork: Artwork }) => {
     useEffect(() => {
         document.body.style.overflow = 'hidden';
         return () => {
@@ -40,7 +34,7 @@ const AuctionDetailPage = ({ artwork = DUMMY_DATA }: { artwork: Artwork }) => {
                         />
                         <Grid>
                             <section>
-                                <img src={DUMMY_DATA.imagePath} />
+                                <img src="" />
                             </section>
                             <ItemDetail />
                         </Grid>
