--- client/pages/exhibition/edit.tsx
@@ -1,12 +1,0 @@
-import Editor from '@components/Exhibition/Editor';
-import React from 'react';
-
-const edit = () => {
-    return (
-        <div>
-            <Editor />
-        </div>
-    );
-};
-
-export default edit;
