--- client/pages/exhibition/editor.tsx
@@ -1,12 +1,0 @@
-import Editor from '@components/Exhibition/Editor';
-import React from 'react';
-
-const editor = () => {
-    return (
-        <div>
-            <Editor></Editor>
-        </div>
-    );
-};
-
-export default editor;
